Clear fdReadDup if the dup'ed fd has been closed
It is normal for acquire to timeout and return None, so to definitely
obtain a token one needs to loop until success:

    jobClient = gnumake_tokenpool.JobClient()
    while True:
        token = jobClient.acquire()
        if token:
	    break

    # do some work

    jobClient.release(token)

However, the timeout handler in acquire() does not clear fdReadDup. So
subsequent call do not dup() a new copy of the fd, and try to read
from the already-closed fd. Effectively, this means that if the
timeout was hit once then it behaves like the timeout hits immediatly
in the subsequent calls.